requirements.txt
datasets==4.0.0
pandas==2.3.1
matplotlib==3.10.5
seaborn==0.13.2
tqdm==4.67.1
ipykernel==6.30.1
openai==1.99.9
huggingface-hub==0.34.4
python-dotenv==1.1.1
